fix(account statement): show description
This commit hides the account label and shows the transaction label by
default.
